+++ src/handler.js
@@ -3,8 +3,8 @@ const crypto = require('crypto');
 require('dotenv').config();
 
 // Function for sending twitter DMs
-const twitter = require('./twitter');
-const cards = require('./cards');
+const twitter = require('./lib/twitter/');
+const cards = require('./lib/cards/');
 
 //twitterUserId = '905237435394560000' // @riskmapus
 twitterUserId = '905602080252977152' //@riskmapbot
